% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R, R/glimpse.R, R/utils.R
\name{trunc_dt}
\alias{trunc_dt}
\alias{gplimpse_dt}
\alias{str_dt}
\alias{dim_desc}
\title{Printing utilities}
\usage{
trunc_dt(x, n = NULL, width = NULL, n_extra = NULL)

gplimpse_dt(x, width = NULL)

str_dt(x, ...)

dim_desc(x)
}
\arguments{
\item{x}{Object to format or print.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than option \code{tibble.print_max}. Otherwise, will print
\code{tibble.print_min} rows.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use \code{getOption("tibble.width")} or (if also \code{NULL})
\code{getOption("width")}; the latter displays only the columns that fit on one
screen. You can also set \code{options(tibble.width = Inf)} to override this
default and always print all columns.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the default,
will print information about at most \code{tibble.max_extra_cols} extra columns.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Printing utilities
}
\keyword{internal}
