% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\alias{subset.prt}
\title{NSE subsetting operations}
\usage{
\method{subset}{prt}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{logical expression indicating elements or rows to keep:
    missing values are taken as false.}

\item{select}{expression, indicating columns to select from a
    data frame.}

\item{drop}{passed on to \code{[} indexing operator.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
NSE subsetting operations
}
