% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetting}
\alias{subsetting}
\alias{[[.prt}
\alias{$.prt}
\alias{[.prt}
\title{Subsetting operations}
\usage{
\method{[[}{prt}(x, i, j, ..., exact = TRUE)

\method{$}{prt}(x, name)

\method{[}{prt}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A \code{prt} object.}

\item{i, j}{Row/column indexes. If \code{j} is omitted, \code{i} is used as column
index.}

\item{...}{Generic compatibility: any further arguments are ignored.}

\item{exact}{Generic compatibility: only the default value of \code{TRUE} is
supported.}

\item{name}{
    A literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).}

\item{drop}{Coerce to a vector if fetching one column via \code{tbl[, j]}.
Default \code{FALSE}, ignored when accessing a column via \code{tbl[j]}.}
}
\description{
Subsetting operations
}
