% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R, R/utils.R
\name{glimpse.prt}
\alias{glimpse.prt}
\alias{str_sum}
\alias{str_sum.prt}
\alias{str_sum.data.frame}
\alias{str.prt}
\alias{tbl_sum.prt}
\title{Get a glimpse of your data}
\usage{
\method{glimpse}{prt}(x, width = NULL, ...)

str_sum(x)

\method{str_sum}{prt}(x)

\method{str_sum}{data.frame}(x)

\method{str}{prt}(object, ...)

\method{tbl_sum}{prt}(x)
}
\arguments{
\item{x}{An object to glimpse at.}

\item{width}{Width of output: defaults to the setting of the option
\code{tibble.width} (if finite) or the width of the console.}

\item{...}{Unused, for extensibility.}

\item{object}{any \R object about which you want to have some
    information.}
}
\description{
Get a glimpse of your data
}
