% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt.R
\name{new_prt}
\alias{new_prt}
\alias{as_prt}
\alias{is_prt}
\alias{dim.prt}
\alias{length.prt}
\alias{n_part}
\alias{part_nrow}
\alias{dimnames.prt}
\alias{names.prt}
\alias{as.data.table.prt}
\alias{as.list.prt}
\alias{as.data.frame.prt}
\alias{as.matrix.prt}
\alias{head.prt}
\alias{tail.prt}
\title{Methods for creating and inspecting prt objects}
\usage{
new_prt(files)

as_prt(tbl, n_chunks = 1L, dir = tempfile())

is_prt(x)

\method{dim}{prt}(x)

\method{length}{prt}(x)

n_part(x)

part_nrow(x)

\method{dimnames}{prt}(x)

\method{names}{prt}(x)

\method{as.data.table}{prt}(x, ...)

\method{as.list}{prt}(x, ...)

\method{as.data.frame}{prt}(x, row.names = NULL, optional = FALSE, ...)

\method{as.matrix}{prt}(x, ...)

\method{head}{prt}(x, n = 6L, ...)

\method{tail}{prt}(x, n = 6L, ...)
}
\arguments{
\item{files}{Character vector of file name(s).}

\item{tbl}{An object inheriting from \code{\link[base:data.frame]{base::data.frame()}}. Requires both a
\code{\link[base:nrow]{base::nrow()}} and \code{\link[base:split]{base::split()}} method.}

\item{n_chunks}{Count variable specifying the number of chunks \code{tbl} is
split into.}

\item{dir}{Directory where the chunked \code{\link[fst:fst]{fst::fst()}} objects reside in.}

\item{x}{A \code{prt} object.}

\item{...}{Generic consistency: additional arguments are ignored and a
warning is issued.}

\item{row.names, optional}{Generic consistency: passing anything other than
the default value issues a warning.}

\item{n}{Count variable indicating the number of rows to return.}
}
\description{
Methods for creating and inspecting prt objects
}
