% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\alias{subset.prt}
\alias{subset_quo}
\title{NSE subsetting operations}
\usage{
\method{subset}{prt}(x, subset, select, part_safe = FALSE, drop = FALSE, ...)

subset_quo(
  x,
  subset = NULL,
  select = NULL,
  part_safe = FALSE,
  env = parent.frame()
)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{logical expression indicating elements or rows to keep:
    missing values are taken as false.}

\item{select}{expression, indicating columns to select from a
    data frame.}

\item{part_safe}{Logical flag indicating whether the \code{subset} expression
can be safely be applied to individual partitions.}

\item{drop}{passed on to \code{[} indexing operator.}

\item{...}{further arguments to be passed to or from other methods.}

\item{env}{The environment in which \code{subset} and \code{select} are evaluated in.
This environment is not applicable for quosures because they have their own
environments.}
}
\description{
NSE subsetting operations
}
