% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{subsetting}
\alias{subsetting}
\alias{[[.prt}
\title{Subsetting operations}
\usage{
\method{[[}{prt}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{A \code{prt} object.}

\item{i, j}{Row/column indexes. If \code{j} is omitted, \code{i} is used as column
index.}

\item{...}{Generic compatibility: any further arguments are ignored.}

\item{exact}{Generic compatibility: only the default value of \code{TRUE} is
supported.}
}
\description{
Subsetting operations
}
